+++ src/ClientConfig_TEST.cc
@@ -18,6 +18,7 @@
 #include <gtest/gtest.h>
 #include <fstream>
 #include <string>
+#include <ignition/common/Console.hh>
 #include <ignition/common/Filesystem.hh>
 #include <ignition/common/Util.hh>
 #include "ignition/fuel_tools/ClientConfig.hh"
@@ -359,6 +360,56 @@ TEST(ServerConfig, APIKey)
 
   config.APIKey("my_other_api_key");
   EXPECT_EQ("my_other_api_key", config.APIKey());
+}
+
+/////////////////////////////////////////////////
+TEST(ClientConfig, AsString)
+{
+  common::Console::SetVerbosity(4);
+  {
+    ClientConfig client;
+    std::string str = "Config path: \nCache location: \nServers:\n";
+    EXPECT_EQ(str, client.AsString());
+  }
+
+  {
+    ServerConfig server;
+    std::string str = "URL: \nLocal name: \nVersion: 1.0\nAPI key: \n";
+    EXPECT_EQ(str, server.AsString());
+  }
+
+  {
+    ServerConfig srv;
+    srv.URL("http://serverurl.com");
+    srv.LocalName("local_name");
+    srv.Version("2.0");
+    srv.APIKey("ABCD");
+
+    auto str = srv.AsString();
+    igndbg << str << std::endl;
+
+    EXPECT_NE(str.find("http://serverurl.com"), std::string::npos);
+    EXPECT_NE(str.find("local_name"), std::string::npos);
+    EXPECT_NE(str.find("2.0"), std::string::npos);
+    EXPECT_NE(str.find("ABCD"), std::string::npos);
+  }
+
+  {
+    ClientConfig client;
+    client.SetConfigPath("config/path");
+    client.CacheLocation("cache/location");
+
+    ServerConfig srv;
+    srv.URL("http://serverurl.com");
+    client.AddServer(srv);
+
+    auto str = client.AsString();
+    igndbg << str << std::endl;
+
+    EXPECT_NE(str.find("config/path"), std::string::npos);
+    EXPECT_NE(str.find("cache/location"), std::string::npos);
+    EXPECT_NE(str.find("http://serverurl.com"), std::string::npos);
+  }
 }
 
 //////////////////////////////////////////////////