+++ Changelog.md
@@ -1,5 +1,8 @@
 ## Ignition Fuel Tools 1.x.x
 
+1. Add some AsString functions
+    * [Pull request 54](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/54)
+
 1. Remove server API version from model unique name
     * [Pull request 47](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/47)